requirements.txt
openai
pyhdb
pyodbc
hana_ml
python-dotenv
pymssql
sqlalchemy-hana